chore

+++ issuetype.go
@@ -2,7 +2,7 @@ package backlog
 
 type IssueType struct {
 	Id           uint64 `json:"id"`
-	ProjectId    int    `json:"projectId"`
+	ProjectId    uint64 `json:"projectId"`
 	Name         string `json:"name"`
 	Color        string `json:"color"`
 	DisplayOrder int    `json:"displayOrder"`
